Move a few data structures down into modules
